+++ powerauth-restful-security-spring/src/main/java/com/wultra/security/powerauth/rest/api/spring/exception/PowerAuthExceptionHandler.java
@@ -30,7 +30,9 @@ import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.method.annotation.HandlerMethodValidationException;
 
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 /**
@@ -211,6 +213,22 @@ public class PowerAuthExceptionHandler {
         String details = ex.getConstraintViolations()
                 .stream()
                 .map(v -> v.getPropertyPath() + ": " + v.getMessage())
+                .collect(Collectors.joining(", "));
+        return new ErrorResponse("ERR_VALIDATION", details);
+    }
+
+    /**
+     * Handle handler method validation errors.
+     * @param ex Exception instance
+     * @return Error response
+     */
+    @ExceptionHandler(HandlerMethodValidationException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public @ResponseBody ErrorResponse handleHandlerMethodValidationException(HandlerMethodValidationException ex) {
+        logger.warn("Handler method validation failed", ex);
+        String details = ex.getAllErrors()
+                .stream()
+                .map(e -> Objects.requireNonNullElseGet(e.getDefaultMessage(), e::toString))
                 .collect(Collectors.joining(", "));
         return new ErrorResponse("ERR_VALIDATION", details);
     }