+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Missing request encryption annotation for activation rename endpoint [(#735)](https://github.com/wultra/powerauth-restful-integration/issues/735)
 - Reject blank activation name in activation rename [(#739)](https://github.com/wultra/powerauth-restful-integration/issues/739)
 
 ## [2.2.0] - 2026-07-20